1. float 타입 -> BigDecimal 타입으로 변경 (Update 쿼리 실행 시, 값의 유실이 발생하기 때문) 2. @DynamicUpdate 제거 및 변경 감지를 적용한 상품 엔티티 전체 필드 Update 쿼리 확인 3. 장바구니에 상품을 담을 때, setScale(2, RoundingMode.CEILING)하고 주문 시, 해당 값 그대로 사용 4. 팀 목록 페이지에 Street, City, Zipcode 출력, remark 미출력하도록 수정 5. 엔티티의 @Setter 제거 및 변경 메서드 추가 6. ItemRepository -> ItemJpaRepository / MemberRepository -> MemberJpaRepository 전체 대체 7. 주문 목록 조회 기능, 주문 목록 페이지, 자세히(주문 내역 조회) 버튼 추가 8. 세션 내 정보가 master, team인 경우에만 상품 목록 페이지에서 담기 버튼이 보이도록 수정 9. 주문 내역에서 주문 ID, 배송 상태가 출력되도록 수정

+++ src/main/java/com/side/portfolio/demo/domain/Cart.java
@@ -3,6 +3,7 @@ package com.side.portfolio.demo.domain;
 import lombok.*;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
@@ -26,13 +27,13 @@ public class Cart {
     private int qty;
 
     @Column(name = "cart_price")
-    private float price;
+    private BigDecimal price;
 
     private LocalDateTime createdDate;
     private LocalDateTime modifiedDate;
 
     //정적 팩토리 메서드
-    public static Cart makeCart(Item item, float itemPrice, int qty) {
+    public static Cart makeCart(Item item, BigDecimal itemPrice, int qty) {
         Cart cart = new Cart();
         cart.setItem(item);
         cart.setPrice(itemPrice);
@@ -41,7 +42,7 @@ public class Cart {
     }
 
     @Builder
-    public Cart(Team team, Item item, float price, int qty,
+    public Cart(Team team, Item item, BigDecimal price, int qty,
                 LocalDateTime createdDate, LocalDateTime modifiedDate) {
         this.team = team;
         this.item = item;
